pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "solodet"
version = "0.1.0"
description = "Ground-to-air RGB drone detection for video sequences"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "torchvision",
    "ultralytics>=8.3",
    "sahi>=0.11",
    "opencv-python-headless>=4.8",
    "supervision>=0.19",
    "pycocotools",
    "pyyaml",
    "pandas",
    "tqdm",
    "gdown",
    "huggingface_hub",
]

[project.optional-dependencies]
dev = [
    "jupyterlab",
    "matplotlib",
]

[tool.hatch.build.targets.wheel]
packages = ["solodet"]
